@@ -524,33 +524,38 @@
 	},
 	"scripts": {
 		"vscode:prepublish": "npm run compile",
-		"package": "vsce package --out F:/DIV/VSCode/extension/market/rech-editor-cobol/rech-editor-cobol.vsix",
-		"compile": "node ./node_modules/vscode/bin/install && npm run tslint && tsc --build --force ./tsconfig.json",
+		"package": "@vscode/vsce package --out F:/DIV/VSCode/extension/market/rech-editor-cobol/rech-editor-cobol.vsix",
+		"compile": "npm run lint && tsc --build --force ./tsconfig.json",
 		"watch": "tsc -watch -p ./",
 		"test": "npm run compile && mocha -r ts-node/register ./src/test/**/*.test.ts",
-		"tslint": "tslint --project tsconfig.json --config tslint.json"
+		"lint": "eslint src --ext .ts",
+		"lint:fix": "eslint src --ext .ts --fix"
 	},
 	"dependencies": {
-		"@types/chai": "4.1.6",
-		"@types/iconv-lite": "0.0.1",
-		"@types/q": "^1.0.0",
-		"chai": "4.2.0",
-		"iconv-lite": "^0.4.24",
-		"q": "^1.4.1",
-		"ts-node": "^7.0.1",
-		"tslint": "^5.14.0",
-		"vscode-languageserver": "^6.1.0-next.9",
-		"vscode-languageclient": "^6.1.0-next.9",
-		"vscode-languageserver-textdocument": "^1.0.1",
-		"rech-ts-commons": "^1.0.0"
+		"@types/chai": "^4.3.0",
+		"@types/iconv-lite": "^0.0.1",
+		"@types/q": "^1.5.0",
+		"chai": "^4.3.0",
+		"iconv-lite": "^0.6.3",
+		"q": "^1.5.1",
+		"rech-ts-commons": "^1.0.0",
+		"ts-node": "^10.9.0",
+		"vscode-languageclient": "^9.0.0",
+		"vscode-languageserver": "^9.0.0",
+		"vscode-languageserver-textdocument": "^1.0.12"
 	},
 	"devDependencies": {
-		"@types/mocha": "^2.2.48",
-		"@types/node": "^8.10.25",
-		"mocha": "^5.2.0",
-		"typescript": "^3.0.0",
-		"vsce": "^1.53.0",
-		"vscode": "^1.1.21"
+		"@types/mocha": "^10.0.0",
+		"@types/node": "^16.18.0",
+		"@types/vscode": "^1.105.0",
+		"@typescript-eslint/eslint-plugin": "^8.46.2",
+		"@typescript-eslint/parser": "^8.46.2",
+		"@vscode/test-electron": "^2.5.2",
+		"@vscode/vsce": "^3.2.0",
+		"eslint": "^8.57.1",
+		"eslint-plugin-deprecation": "^3.0.0",
+		"mocha": "^11.7.4",
+		"typescript": "^5.0.0"
 	},
 	"__metadata": {
 		"id": "ab0c1faa-36ab-4b7d-ae3f-d8b9456b89fd",
