@@ -15,9 +15,9 @@ export class MoveInverter {
      * @param buffer buffer with move commands to be inverted
      */
     public invertOperators(buffer: string): string {
-        const regex = /^( +)move\s+([\w\-]+)([\(\w\-\)\, ]*)\s+to\s+([\w\-]+)(\(\w+\-\w+\,\s+\w+\-\w+\)|\(\w+\-\w+\))*(.*$)/gmi;
+        const regex = /^( +)move\s+([\w-]+)([()\w-, ]*)\s+to\s+([\w-]+)(\(\w+-\w+,\s+\w+-\w+\)|\(\w+-\w+\))*(.*$)/gmi;
         const replacedBuffer = buffer.replace(regex, "$1move $4$5 to $2$3$6");
         return replacedBuffer;
     }
 
-};
+}
