@@ -1,4 +1,5 @@
 "use babel";
+import * as os from "os";
 import { Diagnostic, DiagnosticRelatedInformation } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { CobolDiagnostic } from "./cobolDiagnostic";
@@ -67,7 +68,7 @@ export class Diagnostician {
       const documentPath = new Path(textDocument.uri);
       Log.get().info("FindErrorsAndWarnings from " + documentPath);
       const text = textDocument.getText();
-      const dir = new File(Path.tmpdir() + Path.sep() + "Diagnostic" + Path.sep() + require("os").userInfo().username + Path.sep());
+      const dir = new File(Path.tmpdir() + Path.sep() + "Diagnostic" + Path.sep() + os.userInfo().username + Path.sep());
       if (!dir.exists()) {
         dir.mkdir();
       }
