@@ -147,7 +147,7 @@ export class GeradorCobol {
     if (isNull(commentContent) || isNull(startComment)) return;
     const startCommentDelimiter = startComment[0];
     const endCommentDelimiter = startCommentDelimiter.split("").reverse().join("").replace(/>/g, "<");
-    const regexCommentContent = /[^\>]*\b.*(\b|\B)/;
+    const regexCommentContent = /[^>]*\b.*(\b|\B)/;
     commentContent = regexCommentContent.exec(commentContent[0].toString().replace(endCommentDelimiter, ""));
     if (isNull(commentContent)) return;
     let comment = commentContent[0].toString().trim();
@@ -248,11 +248,11 @@ export class GeradorCobol {
     let dots = "";
     const missingDotsNumber = Colunas.COLUNA_FIM - lineText.length;
     if (missingDotsNumber > 0) {
-      for (var i = 1; i <= missingDotsNumber; i++) {
+      for (let i = 1; i <= missingDotsNumber; i++) {
         dots = dots.concat(".");
       }
       await this.editor.setCurrentLine(lineText + dots);
     }
   }
 
-};
+}
