@@ -80,7 +80,7 @@ export class VariableUtils {
         for (let i = line; i >= 0; i--) {
             const currentLine = lines[i];
             let match;
-            match = /^(.*)section[\.\,]?\s*$/i.exec(currentLine)
+            match = /^(.*)section[.,]?\s*$/i.exec(currentLine)
             if (!section && match) {
                 section = <"working-storage" | "linkage" | "file"> match[1];
                 break;
@@ -100,17 +100,17 @@ export class VariableUtils {
         for (let i = line; i >= 0; i--) {
             const currentLine = lines[i];
             let match;
-            match = /^ *(repository|factory|object)[\.\,]?\s*$/i.exec(currentLine)
+            match = /^ *(repository|factory|object)[.,]?\s*$/i.exec(currentLine)
             if (match) {
                 section = match[1].toLowerCase();
                 break;
             }
-            match = /^ *method-id[\.]?\s(.+)[\.]?\s*$/i.exec(currentLine)
+            match = /^ *method-id[.]?\s(.+)[.]?\s*$/i.exec(currentLine)
             if (match) {
                 section = match[1].split(" ")[0];
                 break;
             }
-            if (/^.*data\s+division[\.\,]?\s*$/i.test(currentLine) || /^ *end method[\.]?/i.test(currentLine)) {
+            if (/^.*data\s+division[.,]?\s*$/i.test(currentLine) || /^ *end method[.]?/i.test(currentLine)) {
                 break;
             }
         }
