@@ -5,35 +5,64 @@ import { CompletionUtils } from "../commons/CompletionUtils";
 /**
  * Class to generate LSP Completion Items for Cobol methods modifications
  */
+export enum MethodModifier {
+    RETURNING = "returning",
+    PUBLIC = "public",
+    PROTECTED = "protected",
+    STATIC = "static",
+    OVERRIDE = "override"
+}
+
 export class MethodModifyersCompletion implements CompletionInterface {
 
-    private modification: string;
+    private modification: MethodModifier | string;
 
-    constructor(modification: string) {
+    constructor(modification: MethodModifier | string) {
         this.modification = modification;
     }
 
     public generate(line: number, _column: number, lines: string[]): Promise<CompletionItem[]> {
         return new Promise((resolve) => {
-            let text = lines[line].trimRight();
-            const match = /(\s*method-id\.\s+)([^\s\.]+)\.?(.*)\.?/.exec(text)
+            const match = /(\s*method-id\.\s+)([^\s.()]+)\.?(.*)\.?/.exec(lines[line].trimEnd())
             if (match == null || match.length < 2) {
                 return [];
             }
-            text = match[1] + match[2]
+            const initialText = match[1] + match[2]
+            let text = "";
+            let terminateText = "";
             if (match.length > 3) {
-                const modifyers = match[3].split(/[\s\.]/g);
+                const finalDotPos = match[3].lastIndexOf(".");
+                const modifyers = match[3].substring(0, finalDotPos).split(/[\s.]/g);
                 for (let index = 0; index < modifyers.length; index++) {
                     const modify = modifyers[index].replace(".", "");
                     if (!this.modification.toUpperCase().startsWith(modify.trim().toUpperCase())) {
-                        text = text + " " + modify;
+                        if (this.verifyTerminatedModifier(modify)) {
+                            terminateText = terminateText + " " + modify;
+                        } else {
+                            if (modify.trim().toLowerCase() == MethodModifier.RETURNING) {
+                                let indexReturning = index;
+                                while (indexReturning < modifyers.length) {
+                                    const returningModifyer = modifyers[indexReturning].replace(".", "");
+                                    text = text + " " + returningModifyer;
+                                    indexReturning++;
+                                }
+                                break;
+                            }
+                            if (text.length > 0) {
+                                text = text + " " + modify;
+                            } else {
+                                text = modify;
+                            }
+                        }
                     }
                 }
             }
-            text = text + " " + this.modification + ".";
+            text = text + " " + this.modification + (this.modification == MethodModifier.RETURNING ? " $1" : "") + terminateText + ".";
+            text = text.replace(/\s+/g, ' ');
+            text = initialText + text;
             resolve(
                 [{
-                    label: this.modification.toUpperCase() + ' to method declaration',
+                    label: 'Complete ' + this.modification.toUpperCase() + ' to method declaration',
                     detail: 'Generates the ' + this.modification.toUpperCase() + ' declaration to current method.',
                     insertText: text,
                     insertTextFormat: InsertTextFormat.Snippet,
@@ -44,6 +73,16 @@ export class MethodModifyersCompletion implements CompletionInterface {
                 }]
             );
         });
+    }
+
+    /**
+     * Verifies if the modifier is a terminated one
+     * @param text
+     * @returns
+     */
+    private verifyTerminatedModifier(text: string): boolean {
+        const terminatedModifiers = [MethodModifier.OVERRIDE, MethodModifier.STATIC, MethodModifier.PROTECTED, MethodModifier.PUBLIC];
+        return terminatedModifiers.includes(text.trim().toLowerCase() as MethodModifier);
     }
 
 }
