@@ -26,7 +26,7 @@ export class PerformHighlight implements HighlightInterface {
         const buffer = BufferSplitter.split(text.getText());
         const currentLineContent = buffer[currentLine];
         // If is pharagraph perform
-        if (/\s*perform.+\s*([\w\-]*)((?:\sthru\s.*)|(?:.+\stimes))?[\.|\,]/.exec(currentLineContent.toLowerCase())) {
+        if (/\s*perform.+\s*([\w-]*)((?:\sthru\s.*)|(?:.+\stimes))?[.|,]/.exec(currentLineContent.toLowerCase())) {
             return new DefaultHighlight().positions(text, word, currentLine, currentCharacter)
         }
         const commandColumn = currentLineContent.length - currentLineContent.trimLeft().length
