@@ -11,7 +11,7 @@ import { ContextType } from '../context/parentContext';
  */
 export class DeclareVariableParser implements SymbolParser {
 
-  private regex = /^\s*declare\s+([\w-]+)\s+as\s+(\S+)?\./i;
+  private regex = /^\s*declare\s+(\w+[\d-]*)\s+as\s+(\S+)?\./i;
   private match: RegExpExecArray | null = null;
 
   /**
@@ -52,7 +52,7 @@ export class DeclareVariableParser implements SymbolParser {
     const varName = this.match[1];
     const detail = this.match[2]?.replace('as', '').trim() || '';
 
-    let kind = SymbolKind.Variable;
+    const kind = SymbolKind.Variable;
 
     const symbol = new DocumentSymbol(
       varName,
